+++ Diamant/src/main/java/com/clt/diamant/gui/SingleDocumentWindow.java
@@ -22,7 +22,6 @@ import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
 import java.awt.event.WindowAdapter;
 import java.awt.event.WindowEvent;
-import java.beans.PropertyChangeEvent;
 import java.beans.PropertyChangeListener;
 import java.io.File;
 import java.lang.reflect.InvocationTargetException;
@@ -49,7 +48,6 @@ import javax.swing.KeyStroke;
 import javax.swing.ScrollPaneConstants;
 import javax.swing.SwingConstants;
 import javax.swing.WindowConstants;
-import javax.swing.event.ChangeEvent;
 import javax.swing.event.ChangeListener;
 
 import com.clt.dialog.client.ConnectDialog;
@@ -992,8 +990,7 @@ public class SingleDocumentWindow<DocType extends SingleDocument>
             p.add(new JLabel(plugin.getName() + ", " + Resources.getString("Version")
                     + " "
                     + plugin.getVersion()), BorderLayout.NORTH);
-            p.add(this.getDocument().getPluginSettings(plugin.getClass())
-                    .createEditor(),
+            p.add(this.getDocument().getPluginSettings(plugin.getClass()).createEditor(),
                   BorderLayout.CENTER);
             jtp.addTab(plugin.getName(), plugin.getIcon(), p);
         }