+++ RecognizerClient/src/main/java/com/clt/speech/recognition/test/Values.java
@@ -1,6 +1,0 @@
-package com.clt.speech.recognition.test;
-
-public interface Values {
-
-    public int[] getValues();
-}