+++ plugins/DialogOS_SphinxPlugin/src/main/java/edu/cmu/lti/dialogos/sphinx/plugin/SphinxNode.java
@@ -25,15 +25,20 @@ public class SphinxNode extends AbstractInputNode {
         return Plugin.getRecognizer();
     }
 
-    @Override 
+    @Override
     public RecognitionExecutor createRecognitionExecutor(com.clt.srgf.Grammar recGrammar) {
         try {
             this.getRecognizer().stopRecognition();
         } catch (SpeechException exn) {
             throw new NodeExecutionException(this, Resources.getString("RecognizerError") + ".", exn);
         }
-        recGrammar.requestRobustness(Boolean.TRUE == getProperty(ENABLE_GARBAGE));
-        return new SphinxRecognitionExecutor(getRecognizer());
+
+        if (getSettings().getSilentMode()) {
+            return new SilentRecognitionExecutor();
+        } else {
+            recGrammar.requestRobustness(Boolean.TRUE == getProperty(ENABLE_GARBAGE));
+            return new SphinxRecognitionExecutor(getRecognizer());
+        }
     }
 
     @Override
@@ -52,12 +57,15 @@ public class SphinxNode extends AbstractInputNode {
         return getSettings().getDefaultLanguage();
     }
 
-    /** retrieve the settings from the dialog graph (which is where they are stored -- not within Plugin!) */
+    /**
+     * retrieve the settings from the dialog graph (which is where they are
+     * stored -- not within Plugin!)
+     */
     private Settings getSettings() {
-        if (getGraph() != null && getGraph().getOwner() != null)
-            return ((Settings) getGraph().getOwner()
-                .getPluginSettings(Plugin.class));
-        else
+        if (getGraph() != null && getGraph().getOwner() != null) {
+            return ((Settings) getGraph().getOwner().getPluginSettings(Plugin.class));
+        } else {
             return null;
+        }
     }
 }